* xlog_ng:
  [core] indent
  [core] feeder stream versioning, refactored remote recovery
  [core] fix wording
  [core, feeder] support for v11 rows over network replication
  [core] fix feeder build
  [core] stop statistic aging if no data sources present
  [box] implicit row tags
  [core, box] implement support for v11 snap
  [core] repair magick lookup
  [core, box] factor out conversion to v11 row format
  [core] panic if recovery failed
  [core, box] unified xlog & snap io, unified robust ondisk format
  [core] add read_u16
  [core] remove dead code

+++ mod/silverbox/box.h
@@ -179,7 +179,7 @@ ENUM(messages, MESSAGES);
 struct box_tuple *index_find(struct index *index, void *key);
 
 struct box_txn *txn_alloc(u32 flags);
-u32 box_dispach(struct box_txn *txn, enum box_mode mode, u32 op, struct tbuf *data);
+u32 box_dispach(struct box_txn *txn, enum box_mode mode, u16 op, struct tbuf *data);
 void tuple_txn_ref(struct box_txn *txn, struct box_tuple *tuple);
 void txn_cleanup(struct box_txn *txn);
